mov is_group* into its own file
