+++ include/hipSYCL/sycl/libkernel/group_functions.hpp
@@ -29,27 +29,8 @@
 #define HIPSYCL_LIBKERNEL_GROUP_FUNCTIONS_HPP
 
 #include "backend.hpp"
-#include "group.hpp"
-#include "sub_group.hpp"
+#include "group_traits.hpp"
 #include <type_traits>
-
-namespace hipsycl {
-namespace sycl {
-
-template<class T>
-struct is_group : public std::false_type {};
-
-template<int Dim> 
-struct is_group<group<Dim>> : public std::true_type {};
-
-template<>
-struct is_group<sub_group> : public std::true_type {};
-
-template<class T>
-inline constexpr bool is_group_v = is_group<T>::value;
-
-}
-}
 
 #ifdef SYCL_DEVICE_ONLY
 #include "generic/hiplike/group_functions.hpp"
