+++ include/hipSYCL/sycl/sycl.hpp
@@ -66,6 +66,7 @@
 #include "libkernel/atomic_ref.hpp"
 #include "libkernel/stream.hpp"
 #include "libkernel/sub_group.hpp"
+#include "libkernel/group_traits.hpp"
 #include "libkernel/memory.hpp"
 #if !HIPSYCL_LIBKERNEL_IS_DEVICE_PASS_SPIRV
  // Not yet supported for SPIR-V
